checkpoint: renames "Range/DiscreteRange.ts" to "Range/Discrete.ts"
